feat(login): restore legacy login look with image using credential auth
